Add timeout to startup script
This solves a problem that you can have when you spin up all replica's
at the same time. They all won't be able to find the master and then
start up as standalone instances.

A better solution might be to set "STARTER" to true on one of the
machines, and all other machines will loop and wait for an IP to become
available.